Tests/LibWeb/Text/expected/resize-scrollable-box.txt

  Item 1Item 2Item 3Item 4Item 5Item 6Item 7Item 8Item 9Item 10Item 11Item 12    scrollLeft (before resize): 1130
scrollLeft (after resize): 830
